requirements.txt
fastapi
uvicorn
pydantic
requests
sqlalchemy
indented_logger
dependency_injector
pyyaml
python-bitcoinrpc
python-dotenv
